started laying out new path snippet architecture.
The new design calls for implementing a Path structure that consists of path snippets. These snippets consist of a page state, a series of element action pairs and then ends with another page state. This should be a much more powerful way to go, however the changes required are significant so this branch will be an ongoing project while other things get done.