tons of improvements
With these changes, each audit is its own record set. Elements are no longer unique and shared between pages and audits. Keeping this uniqueness instead of duplicating records led to a lot of brittleness in the system. This also allowed me to improve speed of execution and the accuracy of audits. Additionally, new users won't struggle to get results because of a database conflict.