Pdf report (#456)
* PDF report generation and tests. 

This report isn't complete. It is missing the appendix and page numbers may not be correct

* Updated the pdf table of contents page numbers

* tons of improvements

With these changes, each audit is its own record set. Elements are no longer unique and shared between pages and audits. Keeping this uniqueness instead of duplicating records led to a lot of brittleness in the system. This also allowed me to improve speed of execution and the accuracy of audits. Additionally, new users won't struggle to get results because of a database conflict.

* PDF improvements and removed design system analysis for now

* removed write call from doc generation method

* updated pdf report

* Removed color palette section from pdf temporarily and fixed type for branding

* Fixed description for branding section